Add progress tracking for worker initialization process
